Create Zomato Restaurant Data Extraction Application
The objective of this project is to develop an application for extracting restaurant data from Zomato. The application will utilize web scraping techniques to capture information such as food images, the number of votes, food names, prices and description. The collected dataset can be utilized for various applications, including restaurant analysis, customer preferences, and pricing trends.